Rename analyze-figma-scope to figma-review-design
- Renamed tool folder, files, and all function names
- Updated API endpoint from /api/analyze-figma-scope to /api/figma-review-design
- Updated MCP tool name from 'analyze-figma-scope' to 'figma-review-design'
- Updated all imports and references in index files
- Updated server/readme.md documentation

The new name better reflects the tool's purpose (reviewing designs and
posting questions) and aligns with the other figma-* tool naming convention.
