+++ server/providers/combined/tools/writing-shell-stories/core-logic.ts
@@ -44,7 +44,7 @@ import {
   groupCommentsIntoThreads,
   formatCommentsForContext,
   type FrameMetadata,
-} from '../../../figma/tools/analyze-figma-scope/figma-comment-utils.js';
+} from '../../../figma/tools/figma-review-design/figma-comment-utils.js';
 import type { ScreenAnnotation } from '../shared/screen-annotation.js';
 import { notesToScreenAnnotations } from './note-text-extractor.js';
