--- server/providers/figma/tools/analyze-figma-scope/index.ts
@@ -1,8 +1,0 @@
-/**
- * Analyze Figma Scope Tool
- *
- * Exports the registration function for the analyze-figma-scope MCP tool.
- * This tool analyzes Figma designs and posts clarifying questions as comments.
- */
-
-export { registerAnalyzeFigmaScopeTool } from './analyze-figma-scope.js';
