+++ server/providers/figma/tools/figma-review-design/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Figma Review Design Tool
+ *
+ * Exports the registration function for the figma-review-design MCP tool.
+ * This tool analyzes Figma designs and posts clarifying questions as comments.
+ */
+
+export { registerFigmaReviewDesignTool } from './figma-review-design.js';
