+++ server/providers/figma/tools/index.ts
@@ -3,7 +3,7 @@ import { registerFigmaGetUserTool } from './figma-get-user.js';
 import { registerFigmaGetMetadataForLayerTool } from './figma-get-metadata-for-layer.js';
 import { registerFigmaGetImageDownloadTool } from './figma-get-image-download.js';
 import { registerFigmaGetLayersForPageTool } from './figma-get-layers-for-page.js';
-import { registerAnalyzeFigmaScopeTool } from './analyze-figma-scope/index.js';
+import { registerFigmaReviewDesignTool } from './figma-review-design/index.js';
 
 /**
  * Register all Figma-specific tools with the MCP server
@@ -26,7 +26,7 @@ export function registerFigmaTools(mcp: McpServer, authContext: any): void {
   registerFigmaGetLayersForPageTool(mcp);
   
   // Figma design analysis and question posting tool
-  registerAnalyzeFigmaScopeTool(mcp);
+  registerFigmaReviewDesignTool(mcp);
   
   console.log('  All Figma tools registered (5 tools)');
 }
